型定義ファイルを再定義
