+++ src/commons/db/db.ts
@@ -1,5 +1,5 @@
 import Dexie, { Table } from "dexie";
-import { budgetType } from "../models/App.model";
+import { budgetType } from "../models";
 
 class BudgetDexie extends Dexie {
   budget!: Table<budgetType>;
