+++ src/components/atoms/Layout.tsx
@@ -1,7 +1,8 @@
 import { Center, Stack } from "@chakra-ui/react";
-import { FC, ReactNode } from "react";
+import { FC } from "react";
+import { LayoutProps } from "../../commons/types";
 
-export const PrimaryLayout: FC<{ children: ReactNode }> = ({ children }) => {
+export const PrimaryLayout: FC<LayoutProps> = ({ children }) => {
   return (
     <Stack minH="100vh" spacing={4}>
       {children}
@@ -9,7 +10,7 @@ export const PrimaryLayout: FC<{ children: ReactNode }> = ({ children }) => {
   );
 };
 
-export const FooterLayout: FC<{ children: ReactNode }> = ({ children }) => {
+export const FooterLayout: FC<LayoutProps> = ({ children }) => {
   return (
     <Center as="footer" bgColor="green.300" py={10}>
       <Stack spacing={4}>{children}</Stack>
@@ -17,7 +18,7 @@ export const FooterLayout: FC<{ children: ReactNode }> = ({ children }) => {
   );
 };
 
-export const MainLayout: FC<{ children: ReactNode }> = ({ children }) => {
+export const MainLayout: FC<LayoutProps> = ({ children }) => {
   return (
     <Stack as="main" flexGrow={1} justifyContent="start" alignItems="center" spacing={8}>
       {children}
@@ -25,7 +26,7 @@ export const MainLayout: FC<{ children: ReactNode }> = ({ children }) => {
   );
 };
 
-export const NavbarLayout: FC<{ children: ReactNode }> = ({ children }) => {
+export const NavbarLayout: FC<LayoutProps> = ({ children }) => {
   return (
     <Center as="footer" bgColor="green.300" py={8}>
       {children}
