+++ src/components/molecules/CustomCheckbox.tsx
@@ -1,6 +1,7 @@
 import { FC } from "react";
+import { DeleteCheckboxProps } from "../../commons/types";
 import { SecondaryCheckbox } from "../atoms/Checkbox";
 
-export const DeleteCheckbox: FC<{ onChange?: () => void }> = (props) => {
+export const DeleteCheckbox: FC<DeleteCheckboxProps> = (props) => {
   return <SecondaryCheckbox {...props} />;
 };
