+++ src/components/molecules/DataGrid.tsx
@@ -1,24 +1,8 @@
-// import { ButtonProps } from "@chakra-ui/react";
 import { FC } from "react";
-import { recordsType } from "../../models/App.model";
 import { PrimaryTable, PrimaryTableBody, PrimaryTableHead } from "../atoms/Table";
-import { CustomButtonProps } from "./CustomButton";
+import { DataGridProps } from "../../commons/types";
 
-export const DataGrid: FC<{
-  columnNames: string[];
-  records: recordsType;
-  frontCheckbox?: {
-    Component: FC<{ onChange?: () => void }>;
-    callback: (arg: number) => void;
-  };
-  backButton?: {
-    Component: FC<CustomButtonProps>;
-    callback: (arg: number) => void;
-  };
-  onNumberInputChange?: (recordIndex: number, fieldIndex: number, e: React.ChangeEvent<HTMLInputElement>) => void;
-  onStringInputChange?: (recordIndex: number, fieldIndex: number, e: React.ChangeEvent<HTMLInputElement>) => void;
-  isDisabled?: boolean;
-}> = ({
+export const DataGrid: FC<DataGridProps> = ({
   columnNames,
   records,
   backButton,
@@ -28,6 +12,7 @@ export const DataGrid: FC<{
   isDisabled = false,
 }) => {
   if (records.length === 0) return <></>;
+
   return (
     <PrimaryTable
       Head={<PrimaryTableHead columnNames={columnNames} />}
