+++ src/components/organisms/BudgetDelete.tsx
@@ -1,13 +1,9 @@
-import { Container, UseDisclosureReturn } from "@chakra-ui/react";
 import { FC } from "react";
+import { Container,  } from "@chakra-ui/react";
 import { DeleteDBPopButton } from "../molecules/CustomButton";
+import { BudgetDeleteProps } from "../../commons/types";
 
-export type BudgetDeleteProps = {
-  onYes?: () => void;
-  onNo?: () => void;
-  deleteDBDisclosure: UseDisclosureReturn;
-  isDisabled?: boolean;
-};
+
 export const BudgetDelete: FC<BudgetDeleteProps> = ({ deleteDBDisclosure, ...props }) => {
   return (
     <Container>
