+++ src/components/organisms/BudgetList.tsx
@@ -1,14 +1,10 @@
-import { Container } from "@chakra-ui/react";
 import { FC } from "react";
-import { BudgetListProps } from "../../Pages/App.hook";
+import { Container } from "@chakra-ui/react";
 import { DetailButton } from "../molecules/CustomButton";
 import { DataGrid } from "../molecules/DataGrid";
+import { BudgetListProps } from "../../commons/types";
 
-export const BudgetList: FC<BudgetListProps & { isDisabled: boolean }> = ({
-  budgetListRecords,
-  onBudgetModalOpen,
-  isDisabled,
-}) => {
+export const BudgetList: FC<BudgetListProps> = ({ budgetListRecords, onBudgetModalOpen, isDisabled }) => {
   return (
     <Container>
       <DataGrid
