+++ src/components/templates/Navbar.tsx
@@ -1,5 +1,5 @@
-import { Heading } from "@chakra-ui/react";
 import { FC } from "react";
+import { Heading } from "@chakra-ui/react";
 import { NavbarLayout } from "../atoms/Layout";
 
 export const Navbar: FC = () => {
